@@ -16,38 +16,34 @@
     "build": "pob-build",
     "build:dev": "pob-build",
     "clean": "rm -Rf docs dist test/node6 coverage",
-    "lint": "eslint --fix -c .eslintrc.yml src/",
+    "lint": "eslint --ext .js,.jsx --fix -c .eslintrc.yml src/",
     "prepublish": "ln -s ../../git-hooks/pre-commit .git/hooks/pre-commit || echo",
-    "preversion": "npm run lint && npm run build && npm run build && pob-repository-check-clean",
+    "preversion": "npm run lint && npm run build && pob-repository-check-clean",
     "release": "pob-repository-check-clean && pob-release",
     "version": "pob-version",
     "watch": "pob-watch",
     "watch:dev": "pob-watch"
   },
   "devDependencies": {
-    "babel-eslint": "^6.0.4",
-    "babel-plugin-defines": "^2.0.0",
-    "babel-plugin-discard-module-references": "^1.0.0",
-    "babel-plugin-react-require": "^2.1.0",
-    "babel-plugin-remove-dead-code": "^1.0.1",
-    "babel-plugin-typecheck": "^3.9.0",
-    "babel-preset-es2015-node5": "^1.2.0",
-    "babel-preset-es2015-node6": "^0.2.0",
-    "babel-preset-es2015-webpack": "^6.4.1",
-    "babel-preset-modern-browsers": "^2.0.0",
-    "babel-preset-react": "^6.5.0",
-    "babel-preset-stage-1": "^6.5.0",
-    "eslint": "^2.12.0",
-    "eslint-config-airbnb": "^9.0.1",
-    "eslint-config-pob": "^6.0.1",
-    "eslint-plugin-import": "^1.8.0",
-    "eslint-plugin-jsx-a11y": "^1.4.1",
-    "eslint-plugin-react": "^5.0.1",
-    "pob-babel": "^0.3.0",
-    "pob-release": "^2.0.3",
-    "react": "^15.0.0"
+    "babel-eslint": "^6.1.2",
+    "babel-preset-es2015": "^6.13.2",
+    "babel-preset-es2015-node6": "^0.3.0",
+    "babel-preset-modern-browsers": "^5.1.0",
+    "babel-preset-modern-browsers-stage-1": "^1.0.0",
+    "babel-preset-react": "^6.11.1",
+    "babel-preset-stage-1": "^6.13.0",
+    "eslint": "^3.3.1",
+    "eslint-config-airbnb": "^10.0.1",
+    "eslint-config-pob": "^6.1.5",
+    "eslint-plugin-babel": "^3.3.0",
+    "eslint-plugin-import": "^1.13.0",
+    "eslint-plugin-jsx-a11y": "^2.1.0",
+    "eslint-plugin-react": "^6.1.1",
+    "pob-babel": "^7.1.1",
+    "pob-release": "^2.1.0",
+    "react": "^15.3.0"
   },
   "peerDependencies": {
-    "react": "^15.0.0"
+    "react": "^15.3.0"
   }
 }
