@@ -17,7 +17,7 @@ function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { de
 function _objectWithoutProperties(obj, keys) { var target = {}; for (var i in obj) { if (keys.indexOf(i) >= 0) continue; if (!Object.prototype.hasOwnProperty.call(obj, i)) continue; target[i] = obj[i]; } return target; }
 
 LinkComponent.propTypes = {
-    to: _react.PropTypes.string.isRequired,
+    to: _react.PropTypes.string,
     params: _react.PropTypes.object,
     children: _react.PropTypes.node
 };
@@ -41,7 +41,8 @@ const Props = function () {
 }();
 
 function LinkComponent(_ref, _ref2) {
-    let to = _ref.to;
+    var _ref$to = _ref.to;
+    let to = _ref$to === undefined ? 'default' : _ref$to;
     let params = _ref.params;
     let children = _ref.children;
 
@@ -55,7 +56,9 @@ function LinkComponent(_ref, _ref2) {
 
     return _react2.default.createElement(
         'a',
-        _extends({ href: ctx.urlGenerator(to, params) }, props),
+        _extends({ href: ctx.urlGenerator(to, params) }, props, {
+            __self: this
+        }),
         children
     );
 }
