@@ -9,7 +9,7 @@ function _objectWithoutProperties(obj, keys) { var target = {}; for (var i in ob
 import { PropTypes } from 'react';
 
 LinkComponent.propTypes = {
-    to: PropTypes.string.isRequired,
+    to: PropTypes.string,
     params: PropTypes.object,
     children: PropTypes.node
 };
@@ -33,7 +33,8 @@ var Props = function () {
 }();
 
 export default function LinkComponent(_ref, _ref2) {
-    var to = _ref.to;
+    var _ref$to = _ref.to;
+    var to = _ref$to === undefined ? 'default' : _ref$to;
     var params = _ref.params;
     var children = _ref.children;
 
@@ -47,7 +48,9 @@ export default function LinkComponent(_ref, _ref2) {
 
     return React.createElement(
         'a',
-        _extends({ href: ctx.urlGenerator(to, params) }, props),
+        _extends({ href: ctx.urlGenerator(to, params) }, props, {
+            __self: this
+        }),
         children
     );
 }
